Initial import of our release notes
Previously, the release notes were living as individual pages on our
blog. At the time we started to ship .NET platform libraries as
individual NuGet packages, that was the solution with the least amount
of overhead.

Since we're now fully embracing open source on GitHub it seems much
simpler to provide these release notes as Markdown pages.