Adding some CLI documents (#131)
Adding a CLI folder with documents that pertain to the CLI tooling (such as known-issues). These need to be moved from the CLI repo since editing them there causes...various consequences that makes maintainance of these documents hard and product-impactful.

So far, only the known issues document is here, and a README.md that links to all other documents where needed.