Specifically include Windows 8 supported version(s)
The way the original is written, it looks like only Windows 7SP1+ and Windows 10 are supported, and Windows 8.x might no longer supported.  Especially as compared to the .NET Core 1.x document, which lists "Windows Client, 7 SP1 - 10".  Added a specific mention of 8.1+, per the linked Windows lifecycle document.  (Note - if Windows 8.x is no longer supported, then it should be specifically added to the "Out of support OS versions" section at the bottom.)