Make wget slightly more verbose
`wget -q` is completeley quiet and doesn't even report any errors.
This can lead to a situation where the subsequent command fails but
the user is not made aware of that issue (like with a 404 error,...).

Therefore I propose `wget -nv` which prevents `wget` from being verbose
but reports errors.