Remove VS preview version number for 2.1.20
Looking at other releases, it looks to be very uncommon (this is the only one
I found) to mention VS preview versions in the vs-version field. It is usually
only mentioned in the vs-support field for the corresponding SDK.

The reason to remove it is that 16.7 Preview 4 is not a valid version "number",
due to the spaces in the label.