Remove VS preview version number 5-preview.8
Looking at other releases, it looks to be very uncommon (this is the only one
I found) to mention VS preview versions in the vs-version field. It is usually
only mentioned in the vs-support field for the corresponding SDK, just as
was done here for the `sdk` (not `sdks[0]`) field.

The reason to remove it is that 16.8 Preview 2 is not a valid version "number",
due to the spaces in the label.